We can use this helper in the Calm tests also
