+++ pipeline/transformer/transformer_calm/src/test/scala/weco/pipeline/transformer/calm/transformers/CalmContributorsTest.scala
@@ -2,14 +2,16 @@ package weco.pipeline.transformer.calm.transformers
 
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
-import weco.catalogue.internal_model.identifiers.{IdState, IdentifierType, SourceIdentifier}
+import weco.catalogue.internal_model.identifiers.IdState
 import weco.catalogue.internal_model.work.{Agent, Contributor}
 import weco.catalogue.source_model.generators.CalmRecordGenerators
+import weco.pipeline.transformer.generators.LabelDerivedIdentifiersGenerators
 
 class CalmContributorsTest
     extends AnyFunSpec
     with Matchers
-    with CalmRecordGenerators {
+    with CalmRecordGenerators
+    with LabelDerivedIdentifiersGenerators {
   it("returns an empty list if there's nothing in 'CreatorName'") {
     val record = createCalmRecord
 
@@ -26,26 +28,14 @@ class CalmContributorsTest
       Contributor(
         id = IdState.Unidentifiable,
         agent = Agent(
-          id = IdState.Identifiable(
-            sourceIdentifier = SourceIdentifier(
-              identifierType = IdentifierType.LabelDerived,
-              value = "gabrielle enthoven",
-              ontologyType = "Agent"
-            )
-          ),
+          id = labelDerivedAgentIdentifier("gabrielle enthoven"),
           label = "Gabrielle Enthoven"
         )
       ),
       Contributor(
         id = IdState.Unidentifiable,
         agent = Agent(
-          id = IdState.Identifiable(
-            sourceIdentifier = SourceIdentifier(
-              identifierType = IdentifierType.LabelDerived,
-              value = "simone berbain",
-              ontologyType = "Agent"
-            )
-          ),
+          id = labelDerivedAgentIdentifier("simone berbain"),
           label = "Simone Berbain"
         )
       ),
