+++ pipeline/transformer/transformer_common/src/test/scala/weco/pipeline/transformer/generators/LabelDerivedIdentifiersGenerators.scala
@@ -1,0 +1,28 @@
+package weco.pipeline.transformer.generators
+
+import weco.catalogue.internal_model.identifiers.IdState.Identifiable
+import weco.catalogue.internal_model.identifiers.{
+  IdState,
+  IdentifierType,
+  SourceIdentifier
+}
+
+trait LabelDerivedIdentifiersGenerators {
+  def labelDerivedAgentIdentifier(value: String): Identifiable =
+    IdState.Identifiable(
+      sourceIdentifier = SourceIdentifier(
+        identifierType = IdentifierType.LabelDerived,
+        value = value,
+        ontologyType = "Agent"
+      )
+    )
+
+  def labelDerivedPersonIdentifier(value: String): Identifiable =
+    IdState.Identifiable(
+      sourceIdentifier = SourceIdentifier(
+        identifierType = IdentifierType.LabelDerived,
+        value = value,
+        ontologyType = "Person"
+      )
+    )
+}
